% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy2clipboard.R
\name{copy2clipboard}
\alias{copy2clipboard}
\title{This function copies a data frame to excel}
\usage{
copy2clipboard(df, sep = "\\t", dec = ".", max.size = (200 * 1000))
}
\description{
This function copies a data frame to excel
}
